grab 下沉 C ：core.request/call/response/multi_request/multi_call 改收 task 名/句柄，在 C 内按"校验(在 grab 前 longjmp)→grab→投递→ungrab"执行并返 grab 成败 boolean；srey.lua 对应 wrapper 变薄、删 _grab_multi_tasks/_ungrab_multi_tasks/_grab_protected_call；lcore 提 _task_handle、_check_multi_names/_grab_multi_names 复用；彻底消除 grab 后 luaL_check* longjmp 漏 ungrab 致 task 引用泄漏、关停死锁 copy=0 兜底：新增 CHECK_COPY_FREE 宏(macro.h)替换 lcore/cmds/kcp/coro_utils 共 17 处 if(!copy)FREE(data)；send_multi 改先取 data 后逐元素校验、每条退出路径释放 copy=0，与 multi_request 一致 审计修复：router 中间件解析移到 c_router:add 之前，避免注册失败残留幽灵 C 路由致 dispatch nil-deref；mongo_parse_startsession 入口初始化 *timeout 修响应缺 timeoutMinutes 时的未初始化读；pgsql_affected_rows 返回 int64+strtoll 修 >2^31 行截断为负；http 首个头部行前导 OWS 统一在字段入口拒绝(obs-fold)防请求走私

+++ lib/protocol/pgsql/pgsql.h
@@ -56,6 +56,6 @@ const char *pgsql_get_db(pgsql_ctx *pg);
 /// </summary>
 /// <param name="pgpack">pgpack_ctx 指针，complete 字段须已填充</param>
 /// <returns>受影响的行数，解析失败时返回 0</returns>
-int32_t pgsql_affected_rows(pgpack_ctx *pgpack);
+int64_t pgsql_affected_rows(pgpack_ctx *pgpack);
 
 #endif//PGSQL_H_
